+++ config/metric_test.go
@@ -6,6 +6,7 @@ package config
 import (
 	"context"
 	"errors"
+	"fmt"
 	"net/url"
 	"reflect"
 	"testing"
@@ -66,7 +67,7 @@ func TestMeterProvider(t *testing.T) {
 							},
 							{
 								Periodic: &PeriodicMetricReader{
-									Exporter: MetricExporter{
+									Exporter: PushMetricExporter{
 										Console: Console{},
 										OTLP:    &OTLPMetric{},
 									},
@@ -122,7 +123,7 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus-no-host",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PullMetricExporter{
 						Prometheus: &Prometheus{},
 					},
 				},
@@ -133,7 +134,7 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus-no-port",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PullMetricExporter{
 						Prometheus: &Prometheus{
 							Host: ptr("localhost"),
 						},
@@ -146,7 +147,7 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PullMetricExporter{
 						Prometheus: &Prometheus{
 							Host:              ptr("localhost"),
 							Port:              ptr(8888),
@@ -167,9 +168,9 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-exporter-invalid-protocol",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol: "http/invalid",
+							Protocol: ptr("http/invalid"),
 						},
 					},
 				},
@@ -180,14 +181,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    "http://localhost:4318",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("http://localhost:4318"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -199,14 +200,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-with-path",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    "http://localhost:4318/path/123",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("http://localhost:4318/path/123"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -218,13 +219,13 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-no-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
+							Protocol:    ptr("grpc/protobuf"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -236,14 +237,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-no-scheme",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -255,14 +256,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-invalid-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    " ",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr(" "),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -271,17 +272,17 @@ func TestReader(t *testing.T) {
 			wantErr: &url.Error{Op: "parse", URL: " ", Err: errors.New("invalid URI for request")},
 		},
 		{
-			name: "periodic/grpc-http-none-compression",
+			name: "periodic/otlp-grpc-none-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -290,17 +291,97 @@ func TestReader(t *testing.T) {
 			wantReader: sdkmetric.NewPeriodicReader(otlpGRPCExporter),
 		},
 		{
+			name: "periodic/otlp-grpc-delta-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("delta"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpGRPCExporter),
+		},
+		{
+			name: "periodic/otlp-grpc-cumulative-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("cumulative"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpGRPCExporter),
+		},
+		{
+			name: "periodic/otlp-grpc-lowmemory-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("lowmemory"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpGRPCExporter),
+		},
+		{
+			name: "periodic/otlp-grpc-invalid-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("invalid"),
+						},
+					},
+				},
+			},
+			wantErr: errors.New("unsupported temporality preference \"invalid\""),
+		},
+		{
 			name: "periodic/otlp-grpc-invalid-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "grpc/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("grpc/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("invalid"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -312,14 +393,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    "http://localhost:4318",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("http://localhost:4318"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -331,14 +412,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-with-path",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    "http://localhost:4318/path/123",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("http://localhost:4318/path/123"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -350,13 +431,13 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-no-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
+							Protocol:    ptr("http/protobuf"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -368,14 +449,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-no-scheme",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -387,14 +468,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-invalid-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    " ",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr(" "),
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -406,14 +487,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-none-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -422,17 +503,97 @@ func TestReader(t *testing.T) {
 			wantReader: sdkmetric.NewPeriodicReader(otlpHTTPExporter),
 		},
 		{
+			name: "periodic/otlp-http-cumulative-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("cumulative"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpHTTPExporter),
+		},
+		{
+			name: "periodic/otlp-http-lowmemory-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("lowmemory"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpHTTPExporter),
+		},
+		{
+			name: "periodic/otlp-http-delta-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("delta"),
+						},
+					},
+				},
+			},
+			wantReader: sdkmetric.NewPeriodicReader(otlpHTTPExporter),
+		},
+		{
+			name: "periodic/otlp-http-invalid-temporality",
+			reader: MetricReader{
+				Periodic: &PeriodicMetricReader{
+					Exporter: PushMetricExporter{
+						OTLP: &OTLPMetric{
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
+							Compression: ptr("none"),
+							Timeout:     ptr(1000),
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
+							},
+							TemporalityPreference: ptr("invalid"),
+						},
+					},
+				},
+			},
+			wantErr: errors.New("unsupported temporality preference \"invalid\""),
+		},
+		{
 			name: "periodic/otlp-http-invalid-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    "http/protobuf",
-							Endpoint:    "localhost:4318",
+							Protocol:    ptr("http/protobuf"),
+							Endpoint:    ptr("localhost:4318"),
 							Compression: ptr("invalid"),
 							Timeout:     ptr(1000),
-							Headers: map[string]string{
-								"test": "test1",
+							Headers: []NameStringValuePair{
+								{Name: "test", Value: ptr("test1")},
 							},
 						},
 					},
@@ -444,7 +605,7 @@ func TestReader(t *testing.T) {
 			name: "periodic/no-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{},
+					Exporter: PushMetricExporter{},
 				},
 			},
 			wantErr: errors.New("no valid metric exporter"),
@@ -453,7 +614,7 @@ func TestReader(t *testing.T) {
 			name: "periodic/console-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						Console: Console{},
 					},
 				},
@@ -466,7 +627,7 @@ func TestReader(t *testing.T) {
 				Periodic: &PeriodicMetricReader{
 					Interval: ptr(30_000),
 					Timeout:  ptr(5_000),
-					Exporter: MetricExporter{
+					Exporter: PushMetricExporter{
 						Console: Console{},
 					},
 				},
@@ -719,7 +880,7 @@ func TestView(t *testing.T) {
 				Stream: &ViewStream{
 					Name:          ptr("new_name"),
 					Description:   ptr("new_description"),
-					AttributeKeys: []string{"foo", "bar"},
+					AttributeKeys: ptr(IncludeExclude{Included: []string{"foo", "bar"}}),
 					Aggregation:   &ViewStreamAggregation{Sum: make(ViewStreamAggregationSum)},
 				},
 			},
@@ -917,35 +1078,104 @@ func TestAggregation(t *testing.T) {
 	}
 }
 
-func TestAttributeFilter(t *testing.T) {
+func TestNewIncludeExcludeFilter(t *testing.T) {
 	testCases := []struct {
 		name          string
-		attributeKeys []string
+		attributeKeys *IncludeExclude
 		wantPass      []string
 		wantFail      []string
 	}{
 		{
 			name:          "empty",
-			attributeKeys: []string{},
-			wantPass:      nil,
-			wantFail:      []string{"foo", "bar"},
+			attributeKeys: nil,
+			wantPass:      []string{"foo", "bar"},
+			wantFail:      nil,
 		},
 		{
-			name:          "filter",
-			attributeKeys: []string{"foo"},
-			wantPass:      []string{"foo"},
-			wantFail:      []string{"bar"},
+			name: "filter-with-include",
+			attributeKeys: ptr(IncludeExclude{
+				Included: []string{"foo"},
+			}),
+			wantPass: []string{"foo"},
+			wantFail: []string{"bar"},
+		},
+		{
+			name: "filter-with-exclude",
+			attributeKeys: ptr(IncludeExclude{
+				Excluded: []string{"foo"},
+			}),
+			wantPass: []string{"bar"},
+			wantFail: []string{"foo"},
+		},
+		{
+			name: "filter-with-include-and-exclude",
+			attributeKeys: ptr(IncludeExclude{
+				Included: []string{"bar"},
+				Excluded: []string{"foo"},
+			}),
+			wantPass: []string{"bar"},
+			wantFail: []string{"foo"},
 		},
 	}
 	for _, tt := range testCases {
 		t.Run(tt.name, func(t *testing.T) {
-			got := attributeFilter(tt.attributeKeys)
+			got, err := newIncludeExcludeFilter(tt.attributeKeys)
+			require.NoError(t, err)
 			for _, pass := range tt.wantPass {
 				require.True(t, got(attribute.KeyValue{Key: attribute.Key(pass), Value: attribute.StringValue("")}))
 			}
 			for _, fail := range tt.wantFail {
 				require.False(t, got(attribute.KeyValue{Key: attribute.Key(fail), Value: attribute.StringValue("")}))
 			}
+		})
+	}
+}
+
+func TestNewIncludeExcludeFilterError(t *testing.T) {
+	_, err := newIncludeExcludeFilter(ptr(IncludeExclude{
+		Included: []string{"foo"},
+		Excluded: []string{"foo"},
+	}))
+	require.Equal(t, fmt.Errorf("attribute cannot be in both include and exclude list: foo"), err)
+}
+
+func TestPrometheusReaderOpts(t *testing.T) {
+	testCases := []struct {
+		name        string
+		cfg         Prometheus
+		wantOptions int
+	}{
+		{
+			name:        "no options",
+			cfg:         Prometheus{},
+			wantOptions: 0,
+		},
+		{
+			name: "all set",
+			cfg: Prometheus{
+				WithoutScopeInfo:           ptr(true),
+				WithoutTypeSuffix:          ptr(true),
+				WithoutUnits:               ptr(true),
+				WithResourceConstantLabels: &IncludeExclude{},
+			},
+			wantOptions: 4,
+		},
+		{
+			name: "all set false",
+			cfg: Prometheus{
+				WithoutScopeInfo:           ptr(false),
+				WithoutTypeSuffix:          ptr(false),
+				WithoutUnits:               ptr(false),
+				WithResourceConstantLabels: &IncludeExclude{},
+			},
+			wantOptions: 1,
+		},
+	}
+	for _, tt := range testCases {
+		t.Run(tt.name, func(t *testing.T) {
+			opts, err := prometheusReaderOpts(&tt.cfg)
+			require.NoError(t, err)
+			require.Len(t, opts, tt.wantOptions)
 		})
 	}
 }