+++ instrumentation/google.golang.org/grpc/otelgrpc/test/interceptor_test.go
@@ -857,6 +857,7 @@ func TestUnaryServerInterceptor(t *testing.T) {
 	for _, check := range serverChecks {
 		name := check.grpcCode.String()
 		t.Run(name, func(t *testing.T) {
+			t.Setenv("OTEL_METRICS_EXEMPLAR_FILTER", "always_off")
 			sr := tracetest.NewSpanRecorder()
 			tp := trace.NewTracerProvider(trace.WithSpanProcessor(sr))